+++ mycodo/functions/base_function.py
@@ -85,7 +85,7 @@ class AbstractFunction(AbstractBaseController):
         return self._set_custom_option(CustomController, self.unique_id, option, value)
 
     def get_custom_option(self, option):
-        return self._get_custom_option(self.function.custom_options, option)
+        return self._get_custom_option(CustomController, self.unique_id, option)
 
     def delete_custom_option(self, option):
         return self._delete_custom_option(CustomController, self.unique_id, option)