+++ mycodo/outputs/base_output.py
@@ -626,7 +626,7 @@ class AbstractOutput(AbstractBaseController):
         return self._set_custom_option(Output, self.unique_id, option, value)
 
     def get_custom_option(self, option):
-        return self._get_custom_option(self.output.custom_options, option)
+        return self._get_custom_option(Output, self.unique_id, option)
 
     def delete_custom_option(self, option):
         return self._delete_custom_option(Output, self.unique_id, option)