+++ CHANGELOG.md
@@ -11,6 +11,8 @@
 ### Bugfixes
 
  - Fix Python Code Input logging Exceptions in the Daemon Log
+ - Fix inability to execute Function Commands
+ - Fix get_custom_option() not returning the latest value
 
 ### Miscellaneous